requirements.txt
asgiref==3.5.0; python_version >= '3.7'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
cryptography==36.0.1; python_version >= '3.6'
django-environ==0.8.1
django-extensions==3.1.5
django-rest-knox==4.2.0
django==4.0.2
djangorestframework==3.13.1
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3'
install==1.3.5
mysqlclient==2.1.0
oauth2==1.9.0.post1
oauthlib==3.2.0; python_version >= '3.6'
pycparser==2.21
pyparsing==3.0.7; python_version >= '3.1'
python-decouple==3.6
pytz==2021.3
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlparse==0.4.2; python_version >= '3.5'
tweepy==4.5.0
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'